### 2022-08-13 #4676 @danielskatz
@ryanmrichard - thanks for your submission.  I'll be the editor for it.

Can you suggest potential reviewers? If so, please mention them here in this thread (without tagging them with an @).

One source of potential reviewers is [this list of people](https://bit.ly/joss-reviewers) who have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
