### 2022-08-13 #4676 @danielskatz
Also, @ryanmrichard, you will need to add more to your README - see https://joss.readthedocs.io/en/latest/review_criteria.html#documentation for what the reviewers will look for.
