### 2022-08-13 #4676 @ryanmrichard
@danielskatz Thanks for your comments. As far as reviewers go:

1. robertodr would be a really good choice. He has done quite a bit of research/work pertaining to  creating CMake infrastructure. Plus he is on the list.
2. ashok-arora says they are comfortable reviewing CMake
3. peanutfun is the lowest reviewer on the list with CMake as a language

I admittedly am only going off what's on the list for the latter two and do not know anything about them.

>Also, @ryanmrichard, you will need to add more to your README - see https://joss.readthedocs.io/en/latest/review_criteria.html#documentation for what the reviewers will look for.

I will go ahead and add the requested information to the README. It should not take too long as all of that information is already present in the documentation. I misread the part of the instructions pertaining to needing that information in the README also. Sorry about that.
