### 2022-08-13 ee8921e
updates readme.rst
