### 2022-08-13 3865513
updates readme.rst (#107)
* updates readme.rst

* Committing license headers

* adds contributing section

Co-authored-by: cmakepp[bot] <cmakepp[bot]@github.com>
