### 2022-08-15 #4676 @robertodr
> wave @robertodr - Would you be willing to review this JOSS submission?

Yes, sure thing!
