### 2022-08-16 #4676 @ryanmrichard
Based on your reviewer list maybe zbeekman, pratikvn, or scivision? They are the only other reviewers who list CMake as one of their programming languages (although I suspect a number of the C/C++ reviewers know CMake even if they do not list it).

Based on conference interactions (namely Supercomputing and DOE's ECP conference) I know a lot of scientific software developers are interested in the functionality provided by CMinx, but unfortunately I do not know of too many other academics actually implementing similar functionality. robertodr was the only academic I could think of who is in this space and I arguably do not have a conflict of interest with.
