### 2022-08-17 #4680 @editorialbot
[REVIEW]: CMinx: A CMake Documentation Generator
**Submitting author:** @ryanmrichard (<a href="http://orcid.org/0000-0003-4235-5179">Ryan M. Richard</a>)
**Repository:** https://github.com/CMakePP/CMinx
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v1.0.12
