### 2022-08-30 #4680 @peanutfun
@ryanmrichard Thanks for suggesting me to review your submission 👋 I'll be reviewing your paper and software over the next days and/or weeks. I guess we can have all detailed discussions in issues I'll eventually raise in the CMinx repository. Once I'm through, I will summarize here.
