### 2022-09-01 b478eea
update workflow to discuss Sphinx setup
