+++ docs/source/full_example.rst
@@ -12,32 +12,208 @@
 .. See the License for the specific language governing permissions and
 .. limitations under the License.
 ..
-################
-Example Workflow
-################
+#############################
+Example CMinx/Sphinx Workflow
+#############################
 
-Here we show a start to finish example of using CMinx. We assume you have a
-CMake file, called :code:`example.cmake`, in your current directory. The
-contents of :code:`example.cmake` are:
+The purpose of this page is to show a start to finish example of how to use
+CMinx and Sphinx to generate documentation. For the purposes of this tutorial
+we assume:
 
-.. include:: ../../tests/examples/example.cmake
+- your project has a root directory ``root/``,
+- ``root/docs/`` contains (or will contain) the source for the documentation,
+- ``root/cmake/`` contains CMake modules, and
+- ``root/cmake/example.cmake`` is the documented CMake source file
+
+Summarily we assume a project structure:
+
+.. code::
+
+   └─ root/
+      ├─ docs/
+      └─ cmake/
+         └─ example.cmake
+
+For the purposes of this tutorial the contents of ``example.cmake`` is not too
+important and we opt for a very basic CMake module:
+
+.. include:: ../../tests/examples/basic_function.cmake
    :literal:
 
-To generate the documentation, we enter our system shell (example assumes
-Bash-like shell on a Unix-like system). Assuming CMinx is in our path and we
-want to generate the documentation in the directory :code:`output`, we run:
+********************
+Step 0: Dependencies
+********************
+
+This tutorial assumes you have already installed:
+
+- Sphinx
+  (`install <https://www.sphinx-doc.org/en/master/usage/installation.html>`__
+  instructions)
+- CMinx (:ref:`install <installing_cminx>` instructions)
+
+and that they are in your path.
+
+**************************************
+Step 1: Initialize the docs/ Directory
+**************************************
+
+This step focuses on setting up the ``docs/`` directory for Sphinx-based
+documentation. If your ``docs/`` directory is already setup feel free to skip
+this step. In the ``docs/`` directory run:
+
 
 .. code:: console
 
-      foo@bar:~$ cminx -o output/ example.cmake
-      Writing RST files to /home/foo/output
-      Writing for file /home/foo/example.cmake
-      Writing RST file /home/foo/output/example.rst
+      foo@bar:root/docs$ sphinx-quickstart
 
-The resultant file :code:`output/example.rst`:
+and answer the questions according to your preferences (we highly recommend
+separating the source and build directory). Afterwards you should have a
+directory structure like (assuming you selected separate source and build
+directories):
 
-.. include:: ../../tests/examples/sphinx/source/example.rst
-   :literal:
+.. code::
 
-Place this file in a Sphinx source directory and add it to your :code:`toctree`
-to render it.
+   └─ root/
+      ├─ docs/
+      |  ├─ build/
+      |  ├─ make.bat
+      |  ├─ Makefile
+      |  └─ source/
+      |     ├─ conf.py
+      |     ├─ index.rst
+      |     ├─ _static/
+      |     └─ _templates/
+      └─ cmake/
+         └─ example.cmake
+
+************************
+Step 2: Update index.rst
+************************
+
+When we run CMinx it will generate a directory filled with API documentation
+for our CMake module. We need to decide where to put that directory. There are
+many possibilities, but we recommend a directory in ``root/docs/source``
+which is used to specifically house the API documentation
+(*e.g.*, ``root/docs/source/cmake_api``).
+
+Regardless of where you choose to put the generated files, the next step is to
+modify ``root/docs/source/index.rst`` so that the documentation Sphinx generates
[…]
