+++ setup.cfg
@@ -1,36 +1,0 @@
-[metadata]
-name = CMinx
-url = https://github.com/CMakePP/CMinx
-license = Apache 2.0
-license_files = LICENSE
-description = Documentation Generator for the CMake Language
-long_description = file: README.rst
-long_description_content_type = text/x-rst
-author = CMakePP
-keywords =
-    CMake
-    Documentaion
-classifiers=
-    Development Status :: 4 - Beta
-    License :: OSI Approved :: Apache Software License
-    Programming Language :: Python :: 3 :: Only
-    Topic :: Documentation
-    Topic :: Software Development :: Documentation
-    Topic :: Software Development :: Build Tools
-
-[options]
-packages = find:
-package_dir =
-    = src
-include_package_data = True
-install_requires =
-    antlr4-python3-runtime==4.7.2
-    confuse>=1.7
-    pathspec>=0.9.0
-
-[options.packages.find]
-where=src
-
-[options.entry_points]
-console_scripts =
-    cminx = cminx:main
