### 2022-09-06 ad9d107
Update paper.md
Co-authored-by: Lukas Riedel <34276446+peanutfun@users.noreply.github.com>
