### 2022-09-06 def16a3
Update paper.md
Co-authored-by: Lukas Riedel <34276446+peanutfun@users.noreply.github.com>
