### 2022-09-06 80c373d
Update paper.bib
Co-authored-by: Lukas Riedel <34276446+peanutfun@users.noreply.github.com>
