+++ paper.bib
@@ -1,9 +1,8 @@
-@misc{frontier,
+@article{frontier,
     author = {Barb Helland},
     title = {At the Frontier: DOE Supercomputing Launches the Exascale Era},
     year = {2022},
     publisher = {United States Department of Energy Office of Science},
-    journal = {Website},
     url = {https://www.energy.gov/science/articles/frontier-doe-supercomputing-launches-exascale-era}
 }
 
