### 2022-09-06 87d1439
Update paper.bib
Co-authored-by: Lukas Riedel <34276446+peanutfun@users.noreply.github.com>
