+++ paper.bib
@@ -19,8 +19,7 @@
     author = {Kitware},
     title = {Sphinx Domain for Modern CMake},
     year = {2022},
-    publisher = {GitHub},
-    journal = {GitHub Repository},
+    howpublished = {{GitHub Repository}},
     url = {https://github.com/scikit-build/moderncmakedomain}
 }
 
