### 2022-09-06 10f280a
Update paper.md
Co-authored-by: Lukas Riedel <34276446+peanutfun@users.noreply.github.com>
