+++ paper.md
@@ -25,10 +25,10 @@ bibliography: paper.bib
 # Summary
 
 This manuscript introduces CMinx, a program for generating
-application programming interface (API) documentation for CMake modules
-(code written in the CMake language). Since most of CMinx's intended audience
-is comprised of C/C++ developers, CMinx is designed to operate similar to
-Doxygen [@doxygen] (the *de facto* C/C++ API documentation tool).
+application programming interface (API) documentation written in the CMake
+language, and CMake modules in particular. Since most of CMinx's intended
+audience is comprised of C/C++ developers, CMinx is designed to operate similar
+to Doxygen [@doxygen] (the *de facto* C/C++ API documentation tool).
 Specifically, developers annotate their CMake source with documentation
 comments (traditional CMake block comments, starting with an extra "`[`"
 character). The documentation comments, written in reST [@rest],
