+++ paper.md
@@ -29,9 +29,9 @@ application programming interface (API) documentation written in the CMake
 language, and CMake modules in particular. Since most of CMinx's intended audience
 is comprised of C/C++ developers, CMinx is designed to operate similar to
 Doxygen [@doxygen], the *de facto* C/C++ API documentation tool.
-Specifically, developers annotate their CMake source with documentation
-comments (traditional CMake block comments, starting with an extra "`[`"
-character). The documentation comments, written in reST [@rest],
+Specifically, developers annotate their CMake source with "documentation"
+comments, which are traditional CMake block comments starting with an extra
+"`[`" character. The documentation comments, written in reST [@rest],
 describe to the reader how the functions, parameters, and variables should be
 used. Running CMinx on the annotated source code generates reST files containing
 the API documentation. The reST files can then be converted into static
