+++ paper.bib
@@ -73,3 +73,35 @@
     url = {https://pandoc.org/},
     year = {2022}
 }
+
+@misc{xSDK,
+  url = {https://arxiv.org/abs/1702.08425},
+  author = {Bartlett, Roscoe and Demeshko, Irina and Gamblin, Todd and
+            Hammond, Glenn and Heroux, Michael and Johnson, Jeffrey and
+            Klinvex, Alicia and Li, Xiaoye and McInnes, Lois Curfman and
+            Moulton, J. David and Osei-Kuffuor, Daniel and Sarich, Jason and
+            Smith, Barry and Willenbring, Jim and Yang, Ulrike Meier},
+  title = {xSDK Foundations: Toward an Extreme-scale Scientific Software
+           Development Kit},
+  publisher = {arXiv},
+  year = {2017},
+}
+
+@Article{snir2014,
+  author   = {Marc Snir and Robert W Wisniewski and Jacob A Abraham and
+              Sarita V Adve and Saurabh Bagchi and Pavan Balaji and
+              Jim Belak and Pradip Bose and Franck Cappello and
+              Bill Carlson and Andrew A Chien and Paul Coteus and
+              Nathan A DeBardeleben and Pedro C Diniz and
+              Christian Engelmann and Mattan Erez and Saverio Fazzari and
+              Al Geist and Rinku Gupta and Fred Johnson and
+              Sriram Krishnamoorthy and Sven Leyffer and Dean Liberty and
+              Subhasish Mitra and Todd Munson and Rob Schreiber and
+              Jon Stearley and Eric Van Hensbergen},
+  journal  = {The International Journal of High Performance Computing
+              Applications},
+  title    = {Addressing failures in exascale computing},
+  year     = {2014},
+  pages    = {129-173},
+  volume   = {28},
+}
