### 2022-09-06 8d02317
update docs for config file
