### 2022-09-06 3ad0524
adds notes on using fetchcontent
