### 2022-09-07 483149a
Fix typo in `variable.rst`
Related to openjournals/joss-reviews#4680
