### 2022-09-07 CMakePP/CMinx#121 @robertodr
Add a custom `macro` Sphinx directive
Not really a problem, but it might improve the generated documentation. The generated reST documentation for both functions and macros uses the `.. function::` Sphinx directive. For macros, CMinx adds a note regarding the scoping rules.

**Describe the solution you'd like**
CMinx could define its own directive for macros, to be used as `.. macro::`. When reading the reST files, this would make the difference between macros and functions slightly more immediate. The note about scoping rules should still be added: I'm guessing it's possible to include that bit in the definition of the custom directive.

The current system works fine as well.

**Additional context**

This is related to the JOSS review process openjournals/joss-reviews#4680
