### 2022-09-07 CMakePP/CMinx#122 @robertodr
Add "Edit on GitHub" button to CMinx documentation webpage
It is currently possible to see the sources for the documentation webpages, but to edit one needs to get to the repo and browse it to find the correct file. This is a bit inconvenient if one just wants to quickly suggest a fix to a typo.

**Describe the solution you'd like**

It should be possible to configure the documentation build process to link back the rendered page to its sources on GitHub.

Current system works, but it could be streamlined.

**Additional context**

This is related to the JOSS review process https://github.com/openjournals/joss-reviews/issues/4680
