### 2022-09-07 #4680 @robertodr
@ryanmrichard I'm also done with my review. Apologies that it took so long! The paper is well written:  I've added some very minor suggestions directly on PR 104. This tool will be quite useful. I find the documentation truly excellent, especially the complete example. I've opened some issues that might help improve the project, but addressing them in not a prerequisite for paper acceptance.

- [ ] Quality of writing: https://github.com/CMakePP/CMinx/pull/104
