### 2022-09-07 0c33623
remove (now) unreachable code.
