### 2022-09-07 CMakePP/CMinx#121 @ryanmrichard
@robertodr Thanks for the suggestion, but I'm inclined to punt on it unless you feel strongly about it. If we do punt on this suggestion, it's something that should be revisited if we ever write a Sphinx plugin for CMinx.

AFAIK the only way to create a custom directive is to make a Sphinx plugin. Admittedly writing a Sphinx plugin for a new directive is pretty simple (see [here](http://www.xavierdupre.fr/blog/2015-06-07_nojs.html) for an example). The bigger problem, IMHO, is that (again AFAIK) there's no way to add this directive without passing the complexity on to the user (assuming we need to write a plugin, then the user now has to install the plugin and register it in the `conf.py` file). The additional complexity is pretty minimal, but for many workflows the benefit from the complexity is also pretty minimal (I'm assuming that the new `macro` directive would basically be reST syntactic sugar around a `function` directive plus the disclaimer. I'm also assuming that in a typical workflow users don't actually look at the generated reST before running Sphinx, meaning they still get the same end result).

@AutonomicPerfectionist if we wanted to make it easier to possibly implement this down the road, how hard would it be to make the directives configurable via the configuration file? Something like:

```.yaml
rst:
  function_directive: "function"
  macro_directive: "function"
```

or (depending on whether Confuse allows it):

```.yaml
rst:
  directives:
    function: "function"
    macro: "macro"
```
(basically factor the `directive` out of the keys).
