### 2022-09-07 CMakePP/CMinx#121 @AutonomicPerfectionist
@ryanmrichard should be pretty simple to add such configuration, I'd have to look again but I'm pretty sure Confuse would allow the second option without much fuss
