### 2022-09-07 CMakePP/CMinx#122 @ryanmrichard
To be honest, I never noticed that we didn't have the "edit on GitHub" button. I thought that was there by default. Nonetheless, #125 (hopefully) fixes this issue. At the very least #125 gets the button, I won't know until it's merged whether the links actually work though. My plan is to keep working on #125 until we get the button and working links. This issue will close in the process, but feel free to reopen if after I'm done with #125 the button isn't working.
