### 2022-09-07 06c9baf
Fix typo in `variable.rst` (#120)
Related to openjournals/joss-reviews#4680

Co-authored-by: Ryan Richard <ryanmrichard1@gmail.com>
