### 2022-09-07 fb2d6b1
Update paper.md
Co-authored-by: Roberto Di Remigio <robertodr@users.noreply.github.com>
