### 2022-09-07 818a9cf
Update paper.md
Co-authored-by: Roberto Di Remigio <robertodr@users.noreply.github.com>
