### 2022-09-07 eb2ba9d
Update paper.md
Co-authored-by: Roberto Di Remigio <robertodr@users.noreply.github.com>
