### 2022-09-07 #4680 @ryanmrichard
@peanutfun and @robertodr First off, thank you both for the excellent suggestions and taking the time to review this. I added both of you to the acknowledgements section of the manuscript.

At this point I believe I have addressed all of the comments in https://github.com/CMakePP/CMinx/pull/104 (which is on both checklists) and I believe that I have addressed all three of @peanutfun's remaining checklist items (two of them are still waiting on @peanutfun to accept the solutions). Please let me know if I missed something and thanks again.
