### 2022-09-08 CMakePP/CMinx#121 @robertodr
OK, I understand that this is technically a bit involved, with very little gain. Feel free to disregard and close the issue.
