### 2022-09-08 CMakePP/CMinx#121 @ryanmrichard
I still think we should implement the configuration file option. We can close this issue after that.
