### 2022-09-12 646507b
Add CTest add_test() processor and documentation type
