### 2022-09-17 f9fb26f
Add tests for documenting CTest tests
