+++ tests/test_samples/ctest_add_test.cmake
@@ -1,0 +1,10 @@
+#[[[
+# This is how to document a CTest test.
+# Note that this is not a CMakeTest,
+# but rather a vanilla CMake add_test() command.
+#]]
+add_test(
+    NAME example_test
+    COMMAND echo hello
+    WORKING_DIRECTORY build/
+)
