### 2022-09-17 025b825
Add additional error checking test for documenting ctest add_test()
