### 2022-09-19 7184b41
Document CTest add_test() (#128)
* update VERSION.txt

* update VERSION.txt

* Add CTest add_test() processor and documentation type

* Add config options for documenting undocumented add_test() invocations

* Delete VERSION.txt

* Add tests for documenting CTest tests

* Add sample and corr RST for ctest add_test

* Add documentation for documenting ctest add_test()

* Add test for incorrect ctest add_test params

* Add additional error checking test for documenting ctest add_test()

* Fix minor comment issue

Co-authored-by: cmakepp[bot] <cmakepp[bot]@github.com>
Co-authored-by: Ryan Richard
[…]
