+++ tests/examples/configs/no_include_undocumented_config.yaml
@@ -40,3 +40,9 @@ input:
   # Whether undocumented test sections should have documentation
   # auto-generated.
   include_undocumented_ct_add_section: false
+
+  # Whether undocumented CTest tests should have documentation
+  # auto-generated. This controls whether all vanilla
+  # CMake add_test() commands should be documented,
+  # it has no relation to CMakeTest tests.
+  include_undocumented_add_test: false
