### 2022-09-24 da4bb22
Add support for detecting kwargs in aggregator
