+++ src/cminx/aggregator.py
@@ -98,7 +98,9 @@ class DocumentationAggregator(CMakeListener):
         function_name = def_params[0].getText()
 
         # Extracts function name and adds the completed function documentation to the 'documented' list
-        self.documented.append(FunctionDocumentation(function_name, docstring, params))
+        doc = FunctionDocumentation(function_name, docstring, params)
+        self.documented.append(doc)
+        self.definition_command_stack.append(doc)
 
     def process_macro(self, ctx: CMakeParser.Command_invocationContext, docstring: str):
         """
@@ -123,7 +125,23 @@ class DocumentationAggregator(CMakeListener):
         macro_name = def_params[0].getText()
 
         # Extracts macro name and adds the completed macro documentation to the 'documented' list
-        self.documented.append(MacroDocumentation(macro_name, docstring, params))
+        doc = MacroDocumentation(macro_name, docstring, params)
+        self.documented.append(doc)
+        self.definition_command_stack.append(doc)
+
+    def process_cmake_parse_arguments(self, ctx: CMakeParser.Command_invocationContext, docstring: str):
+        """
+        Determines whether a documented function or macro uses *args or *kwargs.
+
+        :param ctx: Documented command context. Constructed by the Antlr4 parser.
+
+        :param docstring: Cleaned docstring.
+        """
+        if len(self.definition_command_stack) > 0:
+            last_element = self.definition_command_stack[-1]
+            t = type(last_element)
+            if t is FunctionDocumentation or t is MacroDocumentation:
+                last_element.params.append("**kwargs")
 
     def process_ct_add_test(self, ctx: CMakeParser.Command_invocationContext, docstring: str):
         """
@@ -454,6 +472,8 @@ class DocumentationAggregator(CMakeListener):
                 self.documented_classes_stack.append(None)
             elif command == "cpp_end_class":
                 self.documented_classes_stack.pop()
+            elif command == "cmake_parse_arguments":
+                self.process_cmake_parse_arguments(ctx, "")
             elif ((command == "function"
                    or command == "macro")
                   and self.documented_awaiting_function_def is not None):
