### 2022-09-24 467c517
Add tests for cmake_parse_arguments
