+++ tests/examples/sphinx/source/example_no_undocumented_diff_header.rst
@@ -34,6 +34,12 @@ examples.example.cmake
    
 
 
+.. function:: func_kwargs(**kwargs)
+
+   This is a function with **kwargs.
+   
+
+
 .. function:: "${MyFunctionName}"()
 
    This is a documented function, but the name
