+++ paper.bib
@@ -104,4 +104,5 @@
   year     = {2014},
   pages    = {129-173},
   volume   = {28},
+  doi = {10.1177/1094342014522573},
 }
