### 2022-09-26 c0ae5ac
Update paper.bib
Co-authored-by: Lukas Riedel <34276446+peanutfun@users.noreply.github.com>
