+++ paper.bib
@@ -1,11 +1,3 @@
-@article{frontier,
-    author = {Barb Helland},
-    title = {At the Frontier: DOE Supercomputing Launches the Exascale Era},
-    year = {2022},
-    publisher = {United States Department of Energy Office of Science},
-    url = {https://www.energy.gov/science/articles/frontier-doe-supercomputing-launches-exascale-era}
-}
-
 @misc{official_solution,
     author = {tjwrona1992},
     title = {What is the proper way to document a cmake module},
