### 2022-09-26 2fc2da9
Update paper.bib
Co-authored-by: Lukas Riedel <34276446+peanutfun@users.noreply.github.com>
