+++ paper.bib
@@ -67,7 +67,7 @@
 }
 
 @misc{xSDK,
-  url = {https://arxiv.org/abs/1702.08425},
+  doi = {10.48550/arXiv.1702.08425},
   author = {Bartlett, Roscoe and Demeshko, Irina and Gamblin, Todd and
             Hammond, Glenn and Heroux, Michael and Johnson, Jeffrey and
             Klinvex, Alicia and Li, Xiaoye and McInnes, Lois Curfman and
