### 2022-09-27 #4680 @robertodr
I confirm CMakePP/CMinx#121 is beyond the scope of this review.
