### 2022-09-27 #4680 @danielskatz
> I confirm [CMakePP/CMinx#121](https://github.com/CMakePP/CMinx/issues/121) is beyond the scope of this review.

And you are happy for this to be accepted?
