### 2022-09-27 #4680 @danielskatz
@ryanmrichard - other than that you've made me wonder about "invaluable" vs "valuable", I just have a few suggested changes: see https://github.com/CMakePP/CMinx/pull/130. Please merge this, or let me know what you disagree with, then we can proceed to the remaining acceptance/publication steps.
