### 2022-09-27 #4680 @danielskatz
The word is up to you...

In either case, at this point, could you:
