### 2022-09-27 6b0d542
Update README.rst
Adds our Zendo DOI.
