+++ .github/.licenserc.yaml
@@ -23,6 +23,7 @@ header:
     - docs/source/.templates
     - docs/source/developer/uml_diagrams/*.drawio
     - docs/Makefile
+    - docs/requirements.txt
     - tests/
     - LICENSE
     - MANIFEST.in
