### 2022-09-27 #4680 @ryanmrichard
The tag is v1.0.12, and here's a [link](https://github.com/CMakePP/CMinx/releases/tag/v1.0.12) to the release.
The DOI is 10.5281/zenodo.7116609 and here's a [link](https://zenodo.org/record/7116609#.YzMId9LMJKs) to the Zenodo page.
