### 2022-09-27 15d82e9
Update README.rst
swapped zenodo DOI for JOSS DOI
