+++ README.rst
@@ -20,8 +20,8 @@ CMinx
 .. image:: https://codecov.io/gh/CMakePP/CMinx/branch/master/graph/badge.svg?token=wbiPq8Gnrs
    :target: https://codecov.io/gh/CMakePP/CMinx
 
-.. image:: https://zenodo.org/badge/204571361.svg
-   :target: https://zenodo.org/badge/latestdoi/204571361
+.. image:: https://joss.theoj.org/papers/10.21105/joss.04680/status.svg
+   :target: https://doi.org/10.21105/joss.04680
 
 Full documentation can be found `here <https://cmakepp.github.io/CMinx/>`__.
 
